Update K EDA.Rmd
I made a classification tree. We would still have to figure out how to put it in shiny and make it possible to select variables, but the framework is there. Also as it turns out, not many of the variables explain first generation, so we may have to do something with sex after all if we chose this option.